requirements.txt
# --- Core Machine Learning ---
tensorflow==2.16.2
keras==3.5.0
optree==0.11.0
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.2
joblib==1.3.2

# --- Visualization ---
matplotlib==3.8.3
seaborn==0.13.2

# --- API Server ---
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==2.6.4
python-multipart==0.0.9

# --- Prophet Forecasting (optional) ---
prophet==1.1.5
cmdstanpy==1.2.0

# --- Utilities ---
tqdm==4.66.2
requests==2.31.0
h5py==3.11.0
python-dotenv==1.0.1
supabase==2.4.1

# --- Database ---
mysql-connector-python==9.0.0

# --- For compatibility with some TF/Keras ops ---
h5py==3.11.0
